Refactoring ProfileImagePicker
